debugged rect. prism considerations
(at least the behaviour is as it used to be when I perform an "old" call [i.e. all the boxes involved are actually cubic, but the code is able to handle separate x/y and z grid considerations]).

still need to fix the fftshifting redundancy, the contaminant power equation update I wrote down yesterday, and actual dirty image stack considerations!